requirements.txt
# 核心依赖
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.0.0
markdown>=3.5.0
openai-whisper>=20231117
anthropic>=0.7.0
numpy>=1.24.0
pydub>=0.25.1
streamlit>=1.28.0